requirements.txt
aioredis==1.3.1
alpaca-trade-api==1.2.3
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
astroid==2.4.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
distlib==0.3.1
django==3.1.5
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.2
filelock==3.0.12
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.6.3
incremental==17.5.0
isort==5.7.0
kombu==5.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
nest-asyncio==1.5.1
numpy==1.19.4
pandas==1.3.0
prompt-toolkit==3.0.19
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.1.0
pylint==2.6.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
pytz==2020.5
redis==3.5.3
requests==2.26.0
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.6
uvicorn==0.14.0
vine==5.0.0
virtualenv==20.2.2
wcwidth==0.2.5
websocket==0.2.1
websocket-client==0.57.0
websockets==9.1
wrapt==1.12.1
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.2.0
